refactor: 수정한 ErrorMessage 컴포넌트 적용
